Move OpenTelemetry to Datadog bridge to oteldd package
