+++ internal/pkg/telemetry/oteldd/httpclient.go
@@ -1,4 +1,4 @@
-package telemetry
+package oteldd
 
 import (
 	"context"
